Added non-static type SolidAddIn and moved all instance code from  AddInIntegration to SolidAddin.
